+++ modules/AWSPowerShell/Cmdlets/Redshift/Basic/New-RSCluster-Cmdlet.cs
@@ -227,8 +227,9 @@ namespace Amazon.PowerShell.Cmdlets.RS
         /// <summary>
         /// <para>
         /// <para>The Elastic IP (EIP) address for the cluster.</para><para>Constraints: The cluster must be provisioned in EC2-VPC and publicly-accessible through
-        /// an Internet gateway. For more information about provisioning clusters in EC2-VPC,
-        /// go to <a href="https://docs.aws.amazon.com/redshift/latest/mgmt/working-with-clusters.html#cluster-platforms">Supported
+        /// an Internet gateway. Don't specify the Elastic IP address for a publicly accessible
+        /// cluster with availability zone relocation turned on. For more information about provisioning
+        /// clusters in EC2-VPC, go to <a href="https://docs.aws.amazon.com/redshift/latest/mgmt/working-with-clusters.html#cluster-platforms">Supported
         /// Platforms to Launch Your Cluster</a> in the Amazon Redshift Cluster Management Guide.</para>
         /// </para>
         /// </summary>